+++ src/resources/B246/retorno/L040/Registro1T.php
@@ -161,7 +161,7 @@ class Registro1T extends Generico1
 	public function inserirDetalhe($linhaTxt)
     {
         while (isset(RetornoAbstract::$lines[RetornoAbstract::$linesCounter]) &&
-            $this->data['codigo_lote'] == abs(substr(RetornoAbstract::$lines[RetornoAbstract::$linesCounter], 3, 4))) {
+            $this->data['codigo_lote'] == abs((int) substr(RetornoAbstract::$lines[RetornoAbstract::$linesCounter], 3, 4))) {
             RetornoAbstract::$linesCounter++;
             if (isset(RetornoAbstract::$lines[RetornoAbstract::$linesCounter]) &&
                 substr(RetornoAbstract::$lines[RetornoAbstract::$linesCounter], 13, 1) == "T") {
